requirements.txt
pandas>=2.0.0
beautifulsoup4>=4.12.0
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
tiktoken>=0.5.0
python-docx>=0.8.11
Pillow>=10.0.0
azure-cognitiveservices-vision-computervision>=0.9.0 
azure-kusto-data
azure-identity 
sklearn
z_ai
